pyproject.toml
[project.urls]
GitHub    = "https://github.com/BrokenSource/Pyaket"
Changelog = "https://pyaket.dev/about/changelog"
Contact   = "https://pyaket.dev/about/contact"
Docs      = "https://pyaket.dev/docs"
Homepage  = "https://pyaket.dev"

[project]
name        = "pyaket"
description = "📦 Easy Python to Fast Executables"
authors     = [{name="Tremeschin"}]
dynamic     = ["keywords"]
version     = "0.10.0"
readme      = ".github/readme.md"
license     = "AGPL-3.0"
requires-python = ">=3.10"
dependencies = [
    "broken-source @ git+https://github.com/BrokenSource/BrokenSource",
    "rustup~=1.28.2",
]

[project.optional-dependencies]
cross = [
    "cargo-zigbuild~=0.20.1",
    "ziglang~=0.15.1",
]

[dependency-groups]
dev = [
    "pyaket[cross]",
]

[project.scripts]
pyaket        = "pyaket.__main__:main"
rustup-init   = "rustup_shims:init"
cargo         = "rustup_shims:cargo"
cargo-clippy  = "rustup_shims:cargo_clippy"
cargo-fmt     = "rustup_shims:cargo_fmt"
cargo-miri    = "rustup_shims:cargo_miri"
clippy-driver = "rustup_shims:clippy_driver"
rls           = "rustup_shims:rls"
rust-analyzer = "rustup_shims:rust_analyzer"
rust-gdb      = "rustup_shims:rust_gdb"
rust-gdbgui   = "rustup_shims:rust_gdbgui"
rust-lldb     = "rustup_shims:rust_lldb"
rustc         = "rustup_shims:rustc"
rustdoc       = "rustup_shims:rustdoc"
rustfmt       = "rustup_shims:rustfmt"

[build-system]
requires = ["hatchling", "hook@git+https://github.com/BrokenSource/Hook/"]
build-backend = "hatchling.build"

[tool.hatch.metadata.hooks.plugin]

[tool.hatch.build.targets.wheel]
packages = ["pyaket"]

[tool.hatch.build.targets.wheel.force-include]
"pyaket/shims.py" = "rustup_shims.py"

[tool.hatch.metadata]
allow-direct-references = true
